+++ ObservatoryFramework/Files/Journal/Other/LaunchSRV.cs
@@ -3,7 +3,7 @@
     public class LaunchSRV : JournalBase
     {
         public string Loadout { get; init; }
-        public int ID { get; init; }
+        public ulong ID { get; init; }
         public bool PlayerControlled { get; init; }
     }
 }
