+++ ObservatoryFramework/Files/Journal/Other/NpcCrewPaidWage.cs
@@ -2,7 +2,7 @@
 {
     public class NpcCrewPaidWage : JournalBase
     {
-        public int NpcCrewId { get; init; }
+        public ulong NpcCrewId { get; init; }
         public string NpcCrewName { get; init; }
         public int Amount { get; init; }
     }
