+++ ObservatoryFramework/Files/ParameterTypes/ShipyardPrice.cs
@@ -2,7 +2,7 @@
 {
     public class ShipyardPrice
     {
-        public int id { get; init; }
+        public ulong id { get; init; }
         public string ShipType { get; init; }
         public string ShipType_Localised { get; init; }
         public int ShipPrice { get; init; }
